Windows msvc: rename dirent.h->msvc_dirent.h. mh_text: fix mimeconf-win and warning

+++ src/internfile/mh_text.cpp
@@ -198,7 +198,7 @@ bool MimeHandlerText::readnext()
             return false;
         }
     } else {
-        m_text = m_alltext.substr(m_offs, m_pagesz);
+        m_text = m_alltext.substr((size_t)m_offs, m_pagesz);
     }
 
     if (m_text.length() == 0) {
